requirements.txt
absl-py==2.1.0
accelerate==1.1.1
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
albucore==0.0.23
albumentations==2.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio
appdirs==1.4.4
argon2-cffi
argon2-cffi-bindings
arrow
asttokens
async-lru
async-timeout==4.0.3
attrs
Babel
beautifulsoup4
bleach
blinker==1.9.0
boto3==1.35.69
botocore==1.35.69
Bottleneck
Brotli
cached-property
cachetools==5.5.0
cbor==1.0.0
certifi
cffi
chardet==5.2.0
charset-normalizer
click==8.1.7
colorama
coloredlogs==15.0.1
comm
contourpy
cryptography==43.0.3
cycler
datasets==3.0.1
debugpy
decorator
defusedxml
dill==0.3.8
docker-pycreds==0.4.0
doclayout_yolo==0.0.3
editdistance==0.8.1
effdet==0.4.1
einops==0.8.1
entrypoints
evaluate==0.4.3
exceptiongroup
executing
fastapi==0.115.2
fastjsonschema
ffmpy==0.4.0
filelock
fire==0.7.0
FlagEmbedding==1.3.4
Flask==3.1.0
flatbuffers==24.3.25
fonttools
fqdn
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
gmpy2
gradio==5.3.0
gradio_client==1.4.2
grpcio==1.68.0
h11
h2
hpack==4.0.0
httpcore
httpx
huggingface-hub==0.26.1
humanfriendly==10.0
hyperframe
idna==3.10
ijson==3.3.0
importlib_metadata
importlib_resources
inscriptis==2.5.3
iopath==0.1.10
ipykernel
ipython
ipywidgets==8.1.5
ir_datasets==0.5.9
isoduration
itsdangerous==2.2.0
jedi
Jinja2
jmespath==1.0.1
joblib==1.4.2
json5
jsonpointer
jsonschema
jsonschema-specifications
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events
jupyter-lsp
jupyter_client
jupyter_core
jupyter_server
jupyter_server_terminals
jupyterlab
jupyterlab_pygments
jupyterlab_server
jupyterlab_widgets==3.0.13
kiwisolver
layoutparser==0.3.4
lxml==5.3.1
lz4==4.4.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe
matplotlib==3.9.1
matplotlib-inline
mdurl==0.1.2
mistune
mpmath
multidict==6.1.0
multiprocess==0.70.16
munkres==1.1.4
nbclient
nbconvert
nbformat
nest_asyncio
networkx
nltk==3.9.1
notebook
notebook_shim
numexpr
numpy==1.26.4
nvidia-ml-py==12.535.161
nvitop==1.3.2
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.16.3
opencv-python==4.9.0.80
opencv-python-headless==4.11.0.86
orjson==3.10.9
overrides
packaging
pandas==2.1.4
pandocfilters
parso
pdf2image==1.17.0
pdfminer==20191125
pdfminer.six==20231228
pdfplumber==0.11.4
peft==0.14.0
pexpect
pickleshare
pillow
pkgutil_resolve_name
platformdirs
ply
portalocker==3.0.0
prometheus_client
prompt_toolkit
propcache==0.2.0
protobuf==4.25.5
psutil
ptyprocess
pure_eval
py-cpuinfo==9.0.0
pyarrow==17.0.0
pybboxes==0.1.6
pycocotools==2.0.8
pycparser
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments
pyparsing
pypdfium2==4.30.0
PyQt5==5.15.9
PyQt5-sip==12.12.2
PySocks
python-dateutil
python-dotenv==1.0.1
python-json-logger
python-multipart==0.0.12
pytz
PyYAML
pyzmq
referencing
regex==2024.9.11
reportlab==4.2.5
requests
rfc3339-validator
rfc3986-validator
rich==13.9.2
rouge_score==0.1.2
rpds-py
ruff==0.7.0
s3transfer==0.10.4
safetensors==0.4.5
sahi==0.11.19
scikit-learn==1.5.2
scipy==1.14.1
seaborn
semantic-version==2.10.0
Send2Trash
sentence-transformers==3.3.1
sentencepiece==0.2.0
sentry-sdk==2.17.0
setproctitle==1.3.3
shapely==2.0.6
shellingham==1.5.4
simsimd
sip
six
smmap==5.0.1
sniffio
soupsieve
stack-data
starlette==0.40.0
stringzilla==3.11.3
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
terminado
terminaltables==3.1.10
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
timm==1.0.11
tinycss2
tokenizers==0.21.0
toml
tomli
tomlkit==0.12.0
torch==2.5.0
torchaudio==2.5.0
torchvision==0.20.0
tornado
tqdm
traitlets
transformers==4.49.0
trec-car-tools==2.6
triton==3.1.0
typer==0.12.5
types-python-dateutil
typing-utils
typing_extensions
tzdata
unicodedata2
unlzw3==0.2.3
uri-template
urllib3
uvicorn==0.32.0
wandb==0.16.6
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
wcwidth
webcolors
webencodings
websocket-client
websockets==12.0
Werkzeug==3.1.3
widgetsnbextension==4.0.13
xxhash==3.5.0
yarl==1.16.0
yolov5==7.0.5
zipp
zlib-state==0.1.9